fix: 네비게이션 메뉴의 이모지를 Icon 컴포넌트로 변경 및 디렉토리 구조 수정 (#36)
