+++ src/components/layout/MobileNavBar/MobileNavBar.tsx
@@ -1,5 +1,6 @@
 import { NavLink } from 'react-router-dom';
-import { NAV_ITEM_MOBILE, URL } from '@/constant';
+import { URL } from '@/constant';
+import { mobileNavItems } from '@/navigation';
 import { ProfileImage } from '@/components';
 import * as S from './MobileNavBar.styles';
 
@@ -8,7 +9,7 @@ const SideNavBar = () => {
     <S.MobileNavBar>
       <S.Navigation>
         <ul>
-          {NAV_ITEM_MOBILE.map((navItem, index) =>
+          {mobileNavItems.map((navItem, index) =>
             navItem.link === URL.PROFILE.link ? (
               <S.NavItem key={`${index}-${navItem.text}`}>
                 <ProfileImage link={URL.PROFILE.link} width="2.8rem" />
