+++ src/constant/index.ts
@@ -1,3 +1,2 @@
 export * from './endPoint';
 export * from './url';
-export * from './navigation';
